src/requirements.txt
# numpy==1.22.4
# pandas==1.5.1
# matplotlib==3.5.3
# Pillow==9.2.0
# plotly==5.9.0
# requests==2.28.1
# librosa==0.8.0
# streamlit~=0.87.0
# # streamlit
# tensorflow-cpu
# # opencv-python-headless==4.6.0.66

librosa==0.9.2
streamlit==0.87.0
tensorflow-cpu==2.11.0
numpy==1.23.4
pandas==1.2.5
matplotlib
Pillow
plotly
requests

